pyproject.toml
[project]
name = "mcp-seedance"
version = "0.1.0"
description = "MCP Server for ByteDance Seedance AI Video Generation via AceDataCloud API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "AceDataCloud", email = "support@acedata.cloud" }
]
maintainers = [
    { name = "AceDataCloud", email = "support@acedata.cloud" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "seedance",
    "bytedance",
    "ai",
    "video",
    "generation",
    "acedata",
    "llm",
    "claude",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
]
release = [
    "build>=1.2.0",
    "twine>=5.0.0",
]
all = [
    "mcp-seedance[dev,test,release]",
]

[project.scripts]
mcp-seedance = "main:main"

[project.urls]
Homepage = "https://github.com/AceDataCloud/mcp-seedance"
Documentation = "https://platform.acedata.cloud/documents/seedance-videos-integration"
Repository = "https://github.com/AceDataCloud/mcp-seedance"
Issues = "https://github.com/AceDataCloud/mcp-seedance/issues"
Changelog = "https://github.com/AceDataCloud/mcp-seedance/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling<1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "tools", "prompts", "main.py"]

[tool.hatch.build.targets.sdist]
include = [
    "core/",
    "tools/",
    "prompts/",
    "tests/",
    "main.py",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    ".env.example",
]

# Mypy Configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

# Coverage Configuration
[tool.coverage.run]
source = ["core", "tools"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
